feat: add clap for command line argument parsing and enhance configuration management
