+++ Cargo.toml
@@ -23,6 +23,7 @@
     ] }
     once_cell = "1.21.3"
     thiserror = "2.0.16"
+clap = { version = "4.5.47", features = ["derive"] }
 
 [[bin]]
     name = "my-http-server"
