+++ cofg.yaml
@@ -1,0 +1,30 @@
+addrs:
+  # like: 127.0.0.1
+  ip: 127.0.0.1
+  # like: 80, 8080
+  port: 8080
+middleware:
+  # enabling NormalizePath
+  normalize_path: true
+  # enabling Compress
+  compress: true
+  logger:
+    # enabling logger
+    enabling: true
+    # logger format
+    format: '%{url}xi %s "%{Referer}i" "%{User-Agent}i"'
+# watch: true
+templating:
+  hot_reload: true
+  value:
+    # - "name:value"
+    # value: bool(true/false),num(i64)
+    # - "name:env:ENV_VALUE"
+toc:
+  # make_toc: true
+  path: index.html
+  ext:
+    - html
+    - md
+    - pdf
+public_path: ./public/
