+++ src/test/parser.rs
@@ -17,7 +17,7 @@ fn md2html_minimal_h1() {
   let mut cfg = cofg::Cofg::default();
   cfg.templating.hot_reload = true; // ensure engine reloads template
 
-  let html = parser::md2html("# h1".to_string(), &cfg, vec![]).expect("md2html ok");
+  let html = parser::md2html("# h1".to_string(), &cfg, vec![]).expect("md2html error");
   assert!(html.contains("&lt;h1&gt;h1&lt;/h1&gt;"), "should contain escaped rendered h1: {html}");
   assert!(html.contains(&format!("v{}", env!("CARGO_PKG_VERSION"))), "has version");
